git-svn-id: file://localhost/home/tolik/svnrepo/tags/latest_tested_snapshot@1784 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ opencv/src/cvaux/cvocttree.cpp
@@ -255,7 +255,7 @@ void OctTree::getPointsWithinSphere( const Point3f& center, float radius, Vector
 
 void OctTree::buildTree( const Vector<Point3f>& points3d, int maxLevels, int minPoints)
 {
-    assert( maxLevels * 8 < MAX_STACK_SIZE );
+    assert( (size_t)maxLevels * 8 < MAX_STACK_SIZE );
     points = points3d;
     this->minPoints = minPoints;
